Reworked environment commands.
